+++ packages/base/src/components/Buttons/Button/Button.stories.tsx
@@ -4,6 +4,7 @@ import type { ButtonProps } from './Button.types';
 import { buttonSizeVar, buttonStyleVar } from './Button.types';
 import { ComponentStage } from '../../../styles/Box';
 import { Stack } from '../../Stack';
+import { colorTokens } from '../../../styles';
 
 export default {
   title: 'COMPONENTS/Buttons/Button',
@@ -20,6 +21,11 @@ export default {
       control: { type: 'radio' },
       description: '버튼 사이즈',
       defaultValue: 'M',
+    },
+    color: {
+      options: Object.keys(colorTokens),
+      control: { type: 'select' },
+      defaultValue: 'primary400',
     },
     onClick: { action: 'clicked' },
   },
@@ -39,7 +45,7 @@ Primary.args = {
   styleVar: 'PRIMARY',
   sizeVar: 'M',
   children: '버튼',
-  disabled: true,
+  disabled: false,
 };
 
 export const Secondary: StoryFn<ButtonProps> = (args) => {
@@ -73,6 +79,7 @@ Solid.args = {
   styleVar: 'SOLID',
   sizeVar: 'M',
   children: '버튼',
+  color: 'ocean100',
   disabled: false,
 };