+++ packages/base/src/components/Buttons/IconButton/IconButton.stories.tsx
@@ -6,6 +6,8 @@ import { iconButtonSizeVar, iconButtonStyleVar } from './IconButton.types';
 import { ComponentStage } from '../../../styles/Box';
 import { Stack } from '../../Stack';
 import { ViewOffIcon } from '@shoplflow/shopl-assets';
+import { colorTokens } from '../../../styles';
+import { Icon } from '../../Icon';
 
 export default {
   title: 'COMPONENTS/Buttons/IconButton',
@@ -23,15 +25,59 @@ export default {
       description: '아이콘 버튼 사이즈',
       defaultValue: 'M',
     },
+    color: {
+      options: Object.keys(colorTokens),
+      control: { type: 'select' },
+      defaultValue: 'primary400',
+    },
     onClick: { action: 'clicked' },
   },
+};
+
+export const Primary: StoryFn<IconButtonProps> = (args) => {
+  return (
+    <Stack width={'100px'}>
+      <ComponentStage>
+        <IconButton {...args}>
+          <Icon iconSource={ViewOffIcon} />
+        </IconButton>
+      </ComponentStage>
+    </Stack>
+  );
+};
+
+Primary.args = {
+  styleVar: 'PRIMARY',
+  sizeVar: 'M',
+  color: 'primary400',
+  disabled: false,
+};
+export const Secondary: StoryFn<IconButtonProps> = (args) => {
+  return (
+    <Stack width={'100px'}>
+      <ComponentStage>
+        <IconButton {...args}>
+          <Icon iconSource={ViewOffIcon} />
+        </IconButton>
+      </ComponentStage>
+    </Stack>
+  );
+};
+
+Secondary.args = {
+  styleVar: 'SECONDARY',
+  sizeVar: 'M',
+  color: 'primary400',
+  disabled: false,
 };
 
 export const Solid: StoryFn<IconButtonProps> = (args) => {
   return (
     <Stack width={'100px'}>
       <ComponentStage>
-        <IconButton {...args} />
+        <IconButton {...args}>
+          <Icon iconSource={ViewOffIcon} />
+        </IconButton>
       </ComponentStage>
     </Stack>
   );
@@ -40,7 +86,7 @@ export const Solid: StoryFn<IconButtonProps> = (args) => {
 Solid.args = {
   styleVar: 'SOLID',
   sizeVar: 'M',
-  iconSource: ViewOffIcon,
+  color: 'red200',
   disabled: false,
 };
 
@@ -48,7 +94,9 @@ export const Ghost: StoryFn<IconButtonProps> = (args) => {
   return (
     <Stack width={'100px'}>
       <ComponentStage>
-        <IconButton {...args} iconSize={'S'} sizeVar={'M'} />
+        <IconButton {...args}>
+          <Icon iconSource={ViewOffIcon} />
+        </IconButton>
       </ComponentStage>
     </Stack>
   );
@@ -57,6 +105,6 @@ export const Ghost: StoryFn<IconButtonProps> = (args) => {
 Ghost.args = {
   styleVar: 'GHOST',
   sizeVar: 'S',
-  iconSource: ViewOffIcon,
+
   disabled: false,
 };