+++ packages/base/src/components/Inputs/index.ts
@@ -1,2 +1,3 @@
-export * from './Input/Input';
+export * from './Input';
+export * from './InputButton';
 export * from './TextArea';