+++ packages/base/src/components/Modal/ModalBody.tsx
@@ -1,8 +1,8 @@
 import React, { useCallback, useMemo } from 'react';
-import { BodyContainer, ModalBodyContent } from './Modal.styled';
+import { BodyContainer, ModalBodyContainerInner, ModalBodyContent } from './Modal.styled';
 import type { ModalBodyProps } from './Modal.types';
-import Scrollbars from 'react-custom-scrollbars-2';
 import { useResizeObserver } from '@shoplflow/utils';
+import { ScrollArea } from '../ScrollArea';
 
 const ModalBody = ({
   children,
@@ -46,25 +46,23 @@ const ModalBody = ({
 
   return (
     <BodyContainer isIncludeHeader={isIncludeHeader} height={setAutoHeightMin()}>
-      <Scrollbars
+      <ScrollArea
         id={`scrollbar`}
-        // autoHeight
+        universal
         autoHeight={!modalContainerHeight}
         autoHeightMin={setAutoHeightMin()}
-        // autoHeightMax={'100%'}
         autoHeightMax={windowHeight > 1200 ? heightOverMaxHeight : heightUnderMaxHeight}
-        autoHide
-        autoHideTimeout={1000}
-        autoHideDuration={200}
         style={{
           height: '100%',
           overflow: 'hidden',
         }}
       >
-        <ModalBodyContent isIncludeHeader={isIncludeHeader} sizeVar={sizeVar}>
-          {children}
-        </ModalBodyContent>
-      </Scrollbars>
+        <ModalBodyContainerInner>
+          <ModalBodyContent isIncludeHeader={isIncludeHeader} sizeVar={sizeVar}>
+            {children}
+          </ModalBodyContent>
+        </ModalBodyContainerInner>
+      </ScrollArea>
     </BodyContainer>
   );
 };