+++ packages/base/src/components/ScrollArea/index.ts
@@ -1,0 +1,3 @@
+export { default as ScrollArea } from './ScrollArea';
+export * from './ScrollArea';
+export * from './ScrollArea.types';