+++ packages/base/src/portal/PopperPortal.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import { createPortal } from 'react-dom';
+
+const PopperPortal = () => {
+  return (
+    <>
+      {createPortal(
+        <div
+          id={'popper-portal-key'}
+          style={{
+            zIndex: 20001,
+          }}
+        />,
+        document.body,
+      )}
+    </>
+  );
+};
+
+export default PopperPortal;